Add files via upload
This Contact Management System is a Python-based application that allows users to add AND delete,  for contacts. The system stores contact information (name, phone number, address,gender) in a local file (JSON/CSV) for persistence. It features a simple command-line interface for easy interaction and is perfect for practicing file handling and basic Python programming.